requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
faiss-cpu
tqdm
requests
python-dotenv
sentence-transformers==2.2.2
groq
